Transferred dynamic freezing logic to core from asr
